unstable branch (UNSTABLE VERSION - NEEDS MORE WORK & TESTING !!!) - refactored libv4l API:   - created capture, query and control interfaces   - moved control/query stuff from old capture_device/v4l-control.[ch] into their own interface   - adjusted all driver probes to conform to the new API   - adjusted test apps (list-caps and test-capture) to use the new API   - fixed a printf-related warning in XREALLOC macro   - added LOG_LEVEL and LOG_SOURCE selection in Makefile   - fixed a bug when using YUV411 palette with V4L1 device - Adjusted v4l4j to changes in the API:   - Added VideoDevice class   - Created Control class JNI code to use libv4l's control interface   - Added DeviceInfo, ImageFormat, InputInfo and TunerInfo classes making use of the query interface. These classes are used to report characteristics of video hardware   - Adjusted examples package     - Complete revamp of WebcamViewer which now includes buttons to act on controls   - Adjusted FrameGrabber class to the new libv4l API   - Created JPEGFrameGrabber class   - Adjusted jpeg.c to allow copy of raw images straight to Java if requested - TODO list: review javadoc review README & co files write Junit testcases (specially multi threaded ones) finish working on Control class so menu-type control are supported, increase() and decrease() work move legacy capture_device->actions->list_cap() to where it belongs
git-svn-id: http://v4l4j.googlecode.com/svn/v4l4j/branches/unstable@53 f47d2976-8228-11dd-9899-c7ba65f7c4c7

+++ libv4l/log.h
@@ -80,23 +80,23 @@
 		do { \
 			var = (type) malloc(size); \
 			if (!var) {fprintf(stderr, "[%s:%d %s] MEMALLOC: OUT OF MEMORY. Cant allocate %lu bytes.\n",\
-					__FILE__, __LINE__, __PRETTY_FUNCTION__, size); fflush(stderr);}\
+					__FILE__, __LINE__, __PRETTY_FUNCTION__, (long unsigned int) size); fflush(stderr);}\
 			else { CLEAR(*var); \
-				dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "MEMALLOC: allocating %lu bytes of type %s for var %s (%p).\n", size, #type, #var, var); } \
+				dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "MEMALLOC: allocating %lu bytes of type %s for var %s (%p).\n", (long unsigned int) size, #type, #var, var); } \
 		} while (0)
 
 #define XREALLOC(var, type, size)	\
 		do { \
 			int should_clear = var == NULL ? 1 : 0;\
 			var = (type) realloc(var, size); \
-			if (!var) {fprintf(stderr, "[%s:%d %s] REALLOC: OUT OF MEMORY. Cant reallocate %d bytes.\n",\
-					__FILE__, __LINE__, __PRETTY_FUNCTION__, size); fflush(stderr);}\
+			if (!var) {fprintf(stderr, "[%s:%d %s] REALLOC: OUT OF MEMORY. Cant reallocate %lu bytes.\n",\
+					__FILE__, __LINE__, __PRETTY_FUNCTION__,(long unsigned int)  size); fflush(stderr);}\
 			else { \
 					if (should_clear) {\
 						CLEAR(*var); \
-						dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "REALLOC: Allocating %d bytes of type %s for var %s (%p).\n", size, #type, #var, var); \
+						dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "REALLOC: Allocating %lu bytes of type %s for var %s (%p).\n", (long unsigned int) size, #type, #var, var); \
 					} else \
-						{dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "REALLOC: re-allocating %d bytes of type %s for var %s (%p).\n", size, #type, #var, var);} \
+						{dprint(LIBV4L_LOG_SOURCE_MEMALLOC, LIBV4L_LOG_LEVEL_ALL, "REALLOC: re-allocating %lu bytes of type %s for var %s (%p).\n", (long unsigned int) size, #type, #var, var);} \
 			}\
 		} while (0)
